Add ThumbnailView & divided property from DetailView
